backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.6.4
pandas
numpy
scikit-learn
shap
lime
openai
anthropic
httpx
pytest
passlib[bcrypt]
pyjwt
cryptography
google-auth
supabase
python-multipart
